requirements.txt
# Core / API
alpaca-trade-api>=3.0.0
requests>=2.28.0
websockets>=10.4
ccxt>=3.0.0

# Queue System
redis>=4.5.0
rq>=1.13.0
celery>=5.2.0

# Data & ML
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
tensorflow>=2.10.0
torch>=1.13.0
ta>=0.10.0
joblib>=1.2.0
vectorbt>=0.24.0

# Backtest / Simulación
backtrader>=1.9.76.123
gym>=0.26.0
stable-baselines3>=1.6.0

# Infra / DevOps
fastapi>=0.88.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-dotenv>=0.21.0
prometheus-client>=0.15.0
sentry-sdk>=1.12.0

# Testing
pytest>=7.2.0
tox>=4.0.0

# Utilidades
structlog>=22.3.0
loguru>=0.6.0
tenacity>=8.2.2
PyJWT>=2.8.0
streamlit>=1.31.0
boto3>=1.28.0
schedule>=1.2.0